Convert lexer to use AST source location types
