+++ src/lib/format.rs
@@ -1,11 +1,12 @@
 // Reference: https://github.com/dfinity/candid/blob/master/rust/candid/src/bindings/candid.rs
 
 use crate::ast::{
-    BinOp, BindSort, Case, Dec, DecField, Delim, Exp, ExpField, Literal, Located, Mut, ObjSort,
-    Pat, PrimType, RelOp, Stab, Type, TypeBind, TypeField, UnOp, Vis, Dec_,
+    BinOp, BindSort, Case, Dec, DecField, Dec_, Delim, Exp, ExpField, Literal, Located, Mut,
+    ObjSort, Pat, PrimType, RelOp, Stab, Type, TypeBind, TypeField, UnOp, Vis,
 };
 use crate::format_utils::*;
-use crate::lexer::{is_keyword, GroupSort, Token, TokenTree};
+use crate::lexer::is_keyword;
+use crate::lexer_types::{GroupSort, Token, TokenTree};
 use pretty::RcDoc;
 
 fn format_(doc: RcDoc, width: usize) -> String {
@@ -498,7 +499,7 @@ fn filter_whitespace(trees: &[TokenTree]) -> Vec<&TokenTree> {
 fn filter_whitespace_<'a>(trees: &'a [TokenTree], results: &mut Vec<&'a TokenTree>) {
     for tt in trees {
         if match tt {
-            TokenTree::Token(Token::Space(_), _) => false,
+            TokenTree::Token(Located(Token::Space(_), _)) => false,
             _ => true,
         } {
             results.push(tt);
@@ -507,28 +508,28 @@ fn filter_whitespace_<'a>(trees: &'a [TokenTree], results: &mut Vec<&'a TokenTre
 }
 
 fn get_space<'a>(a: &'a TokenTree, b: &'a TokenTree) -> RcDoc<'a> {
-    use crate::lexer::Token::*;
+    use crate::lexer_types::Token::*;
     use GroupSort::*;
     use TokenTree::*;
     match (a, b) {
         // TODO: refactor these rules to a text-based configuration file
-        (Token(Space(_), _), _) | (_, Token(Space(_), _)) => nil(),
-        (Token(MultiLineSpace(_), _), _) | (_, Token(MultiLineSpace(_), _)) => nil(),
-        (Token(Ident(s), _), Group(_, g, _))
+        (Token(Located(Space(_), _)), _) | (_, Token(Located(Space(_), _))) => nil(),
+        (Token(Located(MultiLineSpace(_), _)), _) | (_, Token(Located(MultiLineSpace(_), _))) => nil(),
+        (Token(Located(Ident(s), _)), Group(_, g, _))
             if !is_keyword(s) && (g == &Paren || g == &Square || g == &Angle) =>
         {
             nil()
         }
-        (Token(Open(_), _), _) | (_, Token(Close(_), _)) => nil(),
-        (_, Token(Delim(_), _)) => nil(),
-        (Token(Delim(_), _), _) => line(),
-        (Token(Dot(_), _), _) => nil(),
-        (Token(Operator(s), _), _) if s.eq("?") => nil(),
-        (_, Token(Operator(s), _)) if s.eq("!") => nil(),
-        (_, Token(Operator(s), _)) if s.starts_with(" ") => nil(),
-        (Token(Operator(s), _), Token(Ident(_), _)) if s.eq("#") => nil(),
-        (_, Token(Dot(_), _)) => wrap_(),
-        (Token(Assign(_), _), _) => wrap(),
+        (Token(Located(Open(_), _)), _) | (_, Token(Located(Close(_), _))) => nil(),
+        (_, Token(Located(Delim(_), _))) => nil(),
+        (Token(Located(Delim(_), _)), _) => line(),
+        (Token(Located(Dot(_), _)), _) => nil(),
+        (Token(Located(Operator(s), _)), _) if s.eq("?") => nil(),
+        (_, Token(Located(Operator(s), _))) if s.eq("!") => nil(),
+        (_, Token(Located(Operator(s), _))) if s.starts_with(" ") => nil(),
+        (Token(Located(Operator(s), _)), Token(Located(Ident(_), _))) if s.eq("#") => nil(),
+        (_, Token(Located(Dot(_), _))) => wrap_(),
+        (Token(Located(Assign(_), _)), _) => wrap(),
         (_, Group(_, Comment, _)) => wrap(),
         (Group(_, Comment, _), _) => line(),
         _ => space(),
@@ -540,7 +541,7 @@ impl ToDoc for TokenTree {
         use GroupSort::*;
         use TokenTree::*;
         match self {
-            Token(t, _) => t.doc(),
+            Token(t) => t.doc(),
             Group(trees, sort, pair) => {
                 let trees = filter_whitespace(trees);
                 let doc = match trees.first() {
@@ -555,7 +556,7 @@ impl ToDoc for TokenTree {
                     }
                 };
                 // let concat = RcDoc::concat(docs);
-                let (open, close) = if let Some(((open, _), (close, _))) = pair {
+                let (open, close) = if let Some((Located(open, _), Located(close, _))) = pair {
                     (&open.data().unwrap()[..], &close.data().unwrap()[..])
                 } else {
                     ("", "") // TODO refactor GroupSort into Option
